Prefer esm over es
`es` is not the best word to identify es modules. It's used by
`material-ui` library to distribute es6+ code. `esm` is a better
alternative which is used by awesome esm loader. In the latest release
rollup started to adrop `esm` word instead of `es`, for now just as an
alias.

`jsnext:main` entry point always was rollup specific and was deprecated
(not officially until we have a real standard) a year ago in favor of
universal `module` field.